rattler_digest

Struct HashingReader

pub struct HashingReader<R, D: Digest> { /* private fields */ }

A simple object that provides a Read implementation that also immediately hashes the bytes read from it. Call HashingReader::finalize to retrieve both the original impl Read object as well as the hash.

If the tokio feature is enabled this object also implements [::tokio::io::AsyncRead] which allows you to use it in an async context as well.

Implementations

impl<R, D: Digest + Default> HashingReader<R, D>

pub fn new(reader: R) -> Self

Constructs a new instance from a reader and a new (empty) hasher.

impl<R, D: Digest> HashingReader<R, D>

pub fn finalize(self) -> (R, Output<D>)

Consumes this instance and returns the original reader and the hash of all bytes read from this instance.